requirements.txt
esptool
python3-wget
